Rename `float` module to `generic`
We will support both integer and float coordinates.
