+++ editor/src/draw/path.rs
@@ -1,7 +1,7 @@
 use i_mesh::path::butt::ButtStrokeBuilder;
 use i_mesh::path::style::StrokeStyle;
-use i_triangle::float::builder::TriangulationBuilder;
-use i_triangle::float::triangulation::Triangulation;
+use i_triangle::generic::builder::TriangulationBuilder;
+use i_triangle::generic::triangulation::Triangulation;
 use i_triangle::i_overlay::i_float::float::point::FloatPoint;
 use i_triangle::i_overlay::i_float::int::point::IntPoint;
 use i_triangle::i_overlay::i_shape::int::path::{IntPath, IntPaths};
