+++ iTriangle/src/generic/custom.rs
@@ -1,6 +1,7 @@
-use crate::float::triangulation::RawTriangulation;
-use crate::int::triangulatable::IntTriangulatable;
+use crate::generic::triangulation::RawTriangulation;
+use crate::int::custom::IntCustomTriangulatable;
 use crate::int::triangulation::RawIntTriangulation;
+use crate::int::validation::Validation;
 use i_key_sort::sort::key::SortKey;
 use i_overlay::i_float::adapter::FloatPointAdapter;
 use i_overlay::i_float::float::compatible::FloatPointCompatible;
@@ -11,52 +12,50 @@ use i_overlay::i_shape::float::adapter::{PathToInt, ShapeToInt, ShapesToInt};
 use i_overlay::i_shape::float::rect::RectInit;
 use i_tree::{Expiration, LayoutNumber};
 
-/// A trait for triangulating float-based geometry with default validation.
+/// A trait for triangulating float geometry with user-defined validation rules.
 ///
-/// Automatically converts the input to integer space, applies validation,
-/// and returns a float-mapped result.
-///
-/// # Implemented For
-/// - `Contour<P>`
-/// - `[Contour<P>]`
-/// - `[Shape<P>]`
-pub trait Triangulatable<P: FloatPointCompatible> {
-    /// Triangulates the shape(s) using the default [`Triangulator`] configuration.
-    ///
-    /// Validation includes contour simplification, direction correction, and area filtering.
-    fn triangulate(&self) -> RawTriangulation<P> {
-        self.triangulate_as::<i32>()
+/// Accepts a custom [`Validation`] object for tuning fill rule, min area, etc.
+pub trait CustomTriangulatable<P: FloatPointCompatible> {
+    /// Performs triangulation using the specified [`Validation`] settings.
+    fn custom_triangulate(&self, validation: Validation<i32>) -> RawTriangulation<P> {
+        self.custom_triangulate_as(validation)
     }
 
-    /// Triangulates the shape(s) using the requested integer coordinate type.
-    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    /// Performs triangulation using the requested integer coordinate type.
+    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey;
 
-    /// Triangulates the shape(s) and inserts the given Steiner points.
-    ///
-    /// Points must lie strictly within the interior of the geometry.
-    fn triangulate_with_steiner_points(&self, points: &[P]) -> RawTriangulation<P> {
-        self.triangulate_with_steiner_points_as::<i32>(points)
+    /// Performs triangulation with Steiner points and a custom [`Validation`] config.
+    fn custom_triangulate_with_steiner_points(
+        &self,
+        points: &[P],
+        validation: Validation<i32>,
+    ) -> RawTriangulation<P> {
+        self.custom_triangulate_with_steiner_points_as(points, validation)
     }
 
-    /// Triangulates the shape(s) with Steiner points using the requested integer coordinate type.
-    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
+    /// Performs triangulation with Steiner points using the requested integer coordinate type.
+    fn custom_triangulate_with_steiner_points_as<I>(
+        &self,
+        points: &[P],
+        validation: Validation<I>,
+    ) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey;
 }
 
-impl<P> Triangulatable<P> for [P]
+impl<P> CustomTriangulatable<P> for Contour<P>
 where
     P: FloatPointCompatible,
 {
-    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_path(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).triangulate();
+            let raw = self.to_int(&adapter).custom_triangulate(validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -66,7 +65,11 @@ where
         }
     }
 
-    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
+    fn custom_triangulate_with_steiner_points_as<I>(
+        &self,
+        points: &[P],
+        validation: Validation<I>,
+    ) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
@@ -75,7 +78,7 @@ where
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .triangulate_with_steiner_points(&float_points);
+                .custom_triangulate_with_steiner_points(&float_points, validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -86,17 +89,17 @@ where
     }
 }
 
-impl<P> Triangulatable<P> for [Contour<P>]
+impl<P> CustomTriangulatable<P> for [Contour<P>]
 where
     P: FloatPointCompatible,
 {
-    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).triangulate();
+            let raw = self.to_int(&adapter).custom_triangulate(validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -106,7 +109,11 @@ where
         }
     }
 
-    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
+    fn custom_triangulate_with_steiner_points_as<I>(
+        &self,
+        points: &[P],
+        validation: Validation<I>,
+    ) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
@@ -115,7 +122,7 @@ where
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .triangulate_with_steiner_points(&float_points);
+                .custom_triangulate_with_steiner_points(&float_points, validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -126,17 +133,17 @@ where
     }
 }
 
-impl<P> Triangulatable<P> for [Shape<P>]
+impl<P> CustomTriangulatable<P> for [Shape<P>]
 where
     P: FloatPointCompatible,
 {
-    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_list_of_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).triangulate();
+            let raw = self.to_int(&adapter).custom_triangulate(validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -146,7 +153,11 @@ where
         }
     }
 
-    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
+    fn custom_triangulate_with_steiner_points_as<I>(
+        &self,
+        points: &[P],
+        validation: Validation<I>,
+    ) -> RawTriangulation<P, I>
     where
         I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
@@ -155,7 +166,7 @@ where
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .triangulate_with_steiner_points(&float_points);
+                .custom_triangulate_with_steiner_points(&float_points, validation);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
