+++ iTriangle/src/generic/mod.rs
