+++ iTriangle/src/generic/triangulatable.rs
@@ -1,6 +1,6 @@
-use crate::float::triangulation::RawTriangulation;
+use crate::generic::triangulation::RawTriangulation;
+use crate::int::triangulatable::IntTriangulatable;
 use crate::int::triangulation::RawIntTriangulation;
-use crate::int::unchecked::IntUncheckedTriangulatable;
 use i_key_sort::sort::key::SortKey;
 use i_overlay::i_float::adapter::FloatPointAdapter;
 use i_overlay::i_float::float::compatible::FloatPointCompatible;
@@ -9,52 +9,54 @@ use i_overlay::i_float::int::number::int::IntNumber;
 use i_overlay::i_shape::base::data::{Contour, Shape};
 use i_overlay::i_shape::float::adapter::{PathToInt, ShapeToInt, ShapesToInt};
 use i_overlay::i_shape::float::rect::RectInit;
-use i_tree::Expiration;
+use i_tree::{Expiration, LayoutNumber};
 
-/// A trait for triangulating already valid float-based geometry.
+/// A trait for triangulating float-based geometry with default validation.
 ///
-/// Skips all validation for performance. Ideal when input is generated programmatically.
+/// Automatically converts the input to integer space, applies validation,
+/// and returns a float-mapped result.
 ///
-/// # Safety Requirements
-/// - Outer contours must be counter-clockwise
-/// - Holes must be clockwise
-/// - Steiner points must lie strictly within the shape
-pub trait UncheckedTriangulatable<P: FloatPointCompatible> {
-    /// Triangulates float geometry without validation or simplification.
-    fn unchecked_triangulate(&self) -> RawTriangulation<P> {
-        self.unchecked_triangulate_as::<i32>()
+/// # Implemented For
+/// - `Contour<P>`
+/// - `[Contour<P>]`
+/// - `[Shape<P>]`
+pub trait Triangulatable<P: FloatPointCompatible> {
+    /// Triangulates the shape(s) using the default [`Triangulator`] configuration.
+    ///
+    /// Validation includes contour simplification, direction correction, and area filtering.
+    fn triangulate(&self) -> RawTriangulation<P> {
+        self.triangulate_as::<i32>()
     }
 
-    /// Triangulates float geometry without validation using the requested integer coordinate type.
-    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    /// Triangulates the shape(s) using the requested integer coordinate type.
+    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey;
+        I: IntNumber + Expiration + LayoutNumber + SortKey;
 
-    /// Same as `unchecked_triangulate`, but inserts user-defined Steiner points.
-    fn unchecked_triangulate_with_steiner_points(&self, points: &[P]) -> RawTriangulation<P> {
-        self.unchecked_triangulate_with_steiner_points_as::<i32>(points)
+    /// Triangulates the shape(s) and inserts the given Steiner points.
+    ///
+    /// Points must lie strictly within the interior of the geometry.
+    fn triangulate_with_steiner_points(&self, points: &[P]) -> RawTriangulation<P> {
+        self.triangulate_with_steiner_points_as::<i32>(points)
     }
 
-    /// Same as `unchecked_triangulate_as`, but inserts user-defined Steiner points.
-    fn unchecked_triangulate_with_steiner_points_as<I>(
-        &self,
-        points: &[P],
-    ) -> RawTriangulation<P, I>
+    /// Triangulates the shape(s) with Steiner points using the requested integer coordinate type.
+    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey;
+        I: IntNumber + Expiration + LayoutNumber + SortKey;
 }
 
-impl<P> UncheckedTriangulatable<P> for [P]
+impl<P> Triangulatable<P> for [P]
 where
     P: FloatPointCompatible,
 {
-    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_path(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).uncheck_triangulate();
+            let raw = self.to_int(&adapter).triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -64,19 +66,16 @@ where
         }
     }
 
-    fn unchecked_triangulate_with_steiner_points_as<I>(
-        &self,
-        points: &[P],
-    ) -> RawTriangulation<P, I>
+    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_path(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .uncheck_triangulate_with_steiner_points(&float_points);
+                .triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -87,17 +86,17 @@ where
     }
 }
 
-impl<P> UncheckedTriangulatable<P> for [Contour<P>]
+impl<P> Triangulatable<P> for [Contour<P>]
 where
     P: FloatPointCompatible,
 {
-    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).uncheck_triangulate();
+            let raw = self.to_int(&adapter).triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -107,19 +106,16 @@ where
         }
     }
 
-    fn unchecked_triangulate_with_steiner_points_as<I>(
-        &self,
-        points: &[P],
-    ) -> RawTriangulation<P, I>
+    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .uncheck_triangulate_with_steiner_points(&float_points);
+                .triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -130,17 +126,17 @@ where
     }
 }
 
-impl<P> UncheckedTriangulatable<P> for [Shape<P>]
+impl<P> Triangulatable<P> for [Shape<P>]
 where
     P: FloatPointCompatible,
 {
-    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
+    fn triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_list_of_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).uncheck_triangulate();
+            let raw = self.to_int(&adapter).triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -150,19 +146,16 @@ where
         }
     }
 
-    fn unchecked_triangulate_with_steiner_points_as<I>(
-        &self,
-        points: &[P],
-    ) -> RawTriangulation<P, I>
+    fn triangulate_with_steiner_points_as<I>(&self, points: &[P]) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + SortKey,
+        I: IntNumber + Expiration + LayoutNumber + SortKey,
     {
         if let Some(rect) = FloatRect::with_list_of_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .uncheck_triangulate_with_steiner_points(&float_points);
+                .triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
