+++ iTriangle/src/generic/triangulator.rs
@@ -1,4 +1,4 @@
-use crate::float::triangulation::Triangulation;
+use crate::generic::triangulation::Triangulation;
 use crate::int::triangulation::{IndexType, IntTriangulation};
 use crate::int::triangulator::IntTriangulator;
 use crate::int::validation::Validation;
