+++ iTriangle/src/generic/unchecked.rs
@@ -1,7 +1,6 @@
-use crate::float::triangulation::RawTriangulation;
-use crate::int::custom::IntCustomTriangulatable;
+use crate::generic::triangulation::RawTriangulation;
 use crate::int::triangulation::RawIntTriangulation;
-use crate::int::validation::Validation;
+use crate::int::unchecked::IntUncheckedTriangulatable;
 use i_key_sort::sort::key::SortKey;
 use i_overlay::i_float::adapter::FloatPointAdapter;
 use i_overlay::i_float::float::compatible::FloatPointCompatible;
@@ -10,52 +9,52 @@ use i_overlay::i_float::int::number::int::IntNumber;
 use i_overlay::i_shape::base::data::{Contour, Shape};
 use i_overlay::i_shape::float::adapter::{PathToInt, ShapeToInt, ShapesToInt};
 use i_overlay::i_shape::float::rect::RectInit;
-use i_tree::{Expiration, LayoutNumber};
+use i_tree::Expiration;
 
-/// A trait for triangulating float geometry with user-defined validation rules.
+/// A trait for triangulating already valid float-based geometry.
 ///
-/// Accepts a custom [`Validation`] object for tuning fill rule, min area, etc.
-pub trait CustomTriangulatable<P: FloatPointCompatible> {
-    /// Performs triangulation using the specified [`Validation`] settings.
-    fn custom_triangulate(&self, validation: Validation<i32>) -> RawTriangulation<P> {
-        self.custom_triangulate_as(validation)
+/// Skips all validation for performance. Ideal when input is generated programmatically.
+///
+/// # Safety Requirements
+/// - Outer contours must be counter-clockwise
+/// - Holes must be clockwise
+/// - Steiner points must lie strictly within the shape
+pub trait UncheckedTriangulatable<P: FloatPointCompatible> {
+    /// Triangulates float geometry without validation or simplification.
+    fn unchecked_triangulate(&self) -> RawTriangulation<P> {
+        self.unchecked_triangulate_as::<i32>()
     }
 
-    /// Performs triangulation using the requested integer coordinate type.
-    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
+    /// Triangulates float geometry without validation using the requested integer coordinate type.
+    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey;
+        I: IntNumber + Expiration + SortKey;
 
-    /// Performs triangulation with Steiner points and a custom [`Validation`] config.
-    fn custom_triangulate_with_steiner_points(
-        &self,
-        points: &[P],
-        validation: Validation<i32>,
-    ) -> RawTriangulation<P> {
-        self.custom_triangulate_with_steiner_points_as(points, validation)
+    /// Same as `unchecked_triangulate`, but inserts user-defined Steiner points.
+    fn unchecked_triangulate_with_steiner_points(&self, points: &[P]) -> RawTriangulation<P> {
+        self.unchecked_triangulate_with_steiner_points_as::<i32>(points)
     }
 
-    /// Performs triangulation with Steiner points using the requested integer coordinate type.
-    fn custom_triangulate_with_steiner_points_as<I>(
+    /// Same as `unchecked_triangulate_as`, but inserts user-defined Steiner points.
+    fn unchecked_triangulate_with_steiner_points_as<I>(
         &self,
         points: &[P],
-        validation: Validation<I>,
     ) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey;
+        I: IntNumber + Expiration + SortKey;
 }
 
-impl<P> CustomTriangulatable<P> for Contour<P>
+impl<P> UncheckedTriangulatable<P> for [P]
 where
     P: FloatPointCompatible,
 {
-    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
+    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_path(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).custom_triangulate(validation);
+            let raw = self.to_int(&adapter).uncheck_triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -65,20 +64,19 @@ where
         }
     }
 
-    fn custom_triangulate_with_steiner_points_as<I>(
+    fn unchecked_triangulate_with_steiner_points_as<I>(
         &self,
         points: &[P],
-        validation: Validation<I>,
     ) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_path(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .custom_triangulate_with_steiner_points(&float_points, validation);
+                .uncheck_triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -89,17 +87,17 @@ where
     }
 }
 
-impl<P> CustomTriangulatable<P> for [Contour<P>]
+impl<P> UncheckedTriangulatable<P> for [Contour<P>]
 where
     P: FloatPointCompatible,
 {
-    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
+    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).custom_triangulate(validation);
+            let raw = self.to_int(&adapter).uncheck_triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -109,20 +107,19 @@ where
         }
     }
 
-    fn custom_triangulate_with_steiner_points_as<I>(
+    fn unchecked_triangulate_with_steiner_points_as<I>(
         &self,
         points: &[P],
-        validation: Validation<I>,
     ) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .custom_triangulate_with_steiner_points(&float_points, validation);
+                .uncheck_triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -133,17 +130,17 @@ where
     }
 }
 
-impl<P> CustomTriangulatable<P> for [Shape<P>]
+impl<P> UncheckedTriangulatable<P> for [Shape<P>]
 where
     P: FloatPointCompatible,
 {
-    fn custom_triangulate_as<I>(&self, validation: Validation<I>) -> RawTriangulation<P, I>
+    fn unchecked_triangulate_as<I>(&self) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_list_of_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
-            let raw = self.to_int(&adapter).custom_triangulate(validation);
+            let raw = self.to_int(&adapter).uncheck_triangulate();
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
@@ -153,20 +150,19 @@ where
         }
     }
 
-    fn custom_triangulate_with_steiner_points_as<I>(
+    fn unchecked_triangulate_with_steiner_points_as<I>(
         &self,
         points: &[P],
-        validation: Validation<I>,
     ) -> RawTriangulation<P, I>
     where
-        I: IntNumber + Expiration + LayoutNumber + SortKey,
+        I: IntNumber + Expiration + SortKey,
     {
         if let Some(rect) = FloatRect::with_list_of_paths(self) {
             let adapter = FloatPointAdapter::<P, I>::new(rect);
             let float_points = points.to_int(&adapter);
             let raw = self
                 .to_int(&adapter)
-                .custom_triangulate_with_steiner_points(&float_points, validation);
+                .uncheck_triangulate_with_steiner_points(&float_points);
             RawTriangulation { raw, adapter }
         } else {
             RawTriangulation {
