+++ iTriangle/src/lib.rs
@@ -2,7 +2,7 @@
 extern crate alloc;
 
 pub mod advanced;
-pub mod float;
+pub mod generic;
 pub mod geom;
 mod index;
 pub mod int;
