+++ iTriangle/tests/float_tests.rs
@@ -7,9 +7,9 @@ mod tests {
     use i_overlay::i_shape::base::data::Contour;
     use i_overlay::i_shape::float::area::Area;
     use i_tree::{Expiration, LayoutNumber};
-    use i_triangle::float::triangulatable::Triangulatable;
-    use i_triangle::float::triangulation::Triangulation;
-    use i_triangle::float::triangulator::Triangulator;
+    use i_triangle::generic::triangulatable::Triangulatable;
+    use i_triangle::generic::triangulation::Triangulation;
+    use i_triangle::generic::triangulator::Triangulator;
     use rand::RngExt;
 
     trait TestInt: IntNumber + Expiration + LayoutNumber + SortKey {}
